Inline step logic into Economics, delete engine/steps/
10 step files (2913 lines) absorbed into Economics objects:
- 4 new: HouseholdIncomeEconomics, DemandEconomics,
  HouseholdFinancialEconomics, PriceEquityEconomics
- 3 inlined: FirmProcessingStep → FirmEconomics,
  BankUpdateStep → BankingEconomics,
  WorldAssemblyStep → WorldAssemblyEconomics
- 3 already replaced: FiscalConstraintStep, LaborDemographicsStep,
  OpenEconomyStep (bridge Output types added)

All Step.Output references removed from Economics objects.
FlowSimulation.computeAll() calls Economics exclusively.
engine/steps/ directory deleted.

Fixes #166
