+++ src/test/scala/com/boombustgroup/amorfati/agents/CentralBankPropertySpec.scala
@@ -68,14 +68,14 @@ class CentralBankPropertySpec extends AnyFlatSpec with Matchers with ScalaCheckP
     forAll(Gen.choose(0.0, 0.25), Gen.choose(0.0, 1e10), Gen.choose(1e6, 1e12)) { (rate: Double, bankBonds: Double, gdp: Double) =>
       val nbp      = Nbp.State(Rate(rate), PLN.Zero, true, PLN.Zero, PLN.Zero, PLN.Zero)
       val qeResult = Nbp.executeQe(nbp, PLN(bankBonds), PLN(gdp))
-      // executeQe returns a request; bond update happens in BankUpdateStep
+      // executeQe returns a request; bond update happens in BankingEconomics
       td.toDouble(qeResult.requestedPurchase + nbp.govBondHoldings) should be <= (td.toDouble(p.monetary.qeMaxGdpShare) * gdp + 1e-6)
     }
 
-  it should "not modify nbpState.govBondHoldings (deferred to BankUpdateStep)" in
+  it should "not modify nbpState.govBondHoldings (deferred to BankingEconomics)" in
     forAll(genNbpState, Gen.choose(0.0, 1e10), Gen.choose(1e6, 1e12)) { (nbp: Nbp.State, bankBonds: Double, gdp: Double) =>
       val qeResult = Nbp.executeQe(nbp, PLN(bankBonds), PLN(gdp))
-      // executeQe returns a request; bond update happens in BankUpdateStep waterfall
+      // executeQe returns a request; bond update happens in BankingEconomics waterfall
       qeResult.nbpState.govBondHoldings shouldBe nbp.govBondHoldings
       qeResult.requestedPurchase should be >= PLN.Zero
       td.toDouble(qeResult.requestedPurchase) should be <= (bankBonds + 1.0)
