+++ src/test/scala/com/boombustgroup/amorfati/engine/economics/DynamicNetworkSpec.scala
@@ -1,4 +1,4 @@
-package com.boombustgroup.amorfati.engine.steps
+package com.boombustgroup.amorfati.engine.economics
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
@@ -12,15 +12,15 @@ class DynamicNetworkSpec extends AnyFlatSpec with Matchers:
   import com.boombustgroup.amorfati.config.SimParams
   given SimParams = SimParams.defaults
 
-  "PriceEquityStep.rewireFirms" should "return unchanged firms when rho=0" in {
+  "PriceEquityEconomics.rewireFirms" should "return unchanged firms when rho=0" in {
     val firms  = mkFirms(20)
-    val result = PriceEquityStep.rewireFirms(firms, 0.0, new Random(42))
+    val result = PriceEquityEconomics.rewireFirms(firms, 0.0, new Random(42))
     result shouldBe theSameInstanceAs(firms)
   }
 
   it should "preserve total firm count" in {
     val firms  = mkFirmsWithBankrupt(20, 5)
-    val result = PriceEquityStep.rewireFirms(firms, 1.0, new Random(42))
+    val result = PriceEquityEconomics.rewireFirms(firms, 1.0, new Random(42))
     result.length shouldBe 20
   }
 
@@ -28,7 +28,7 @@ class DynamicNetworkSpec extends AnyFlatSpec with Matchers:
     val firms = mkFirmsWithBankrupt(20, 5)
     firms.count(!Firm.isAlive(_)) shouldBe 5
 
-    val result = PriceEquityStep.rewireFirms(firms, 1.0, new Random(42))
+    val result = PriceEquityEconomics.rewireFirms(firms, 1.0, new Random(42))
     // All bankrupt firms should be replaced
     result.count(!Firm.isAlive(_)) shouldBe 0
     // Replaced firms should be Traditional
@@ -37,26 +37,26 @@ class DynamicNetworkSpec extends AnyFlatSpec with Matchers:
 
   it should "give new firms neighbors" in {
     val firms  = mkFirmsWithBankrupt(30, 3)
-    val result = PriceEquityStep.rewireFirms(firms, 1.0, new Random(42))
+    val result = PriceEquityEconomics.rewireFirms(firms, 1.0, new Random(42))
     // Replaced firms (indices 0-2) should have neighbors
     for i <- 0 until 3 do result(i).neighbors.length should be > 0
   }
 
   it should "return unchanged when no bankrupt firms exist" in {
     val firms  = mkFirms(20)
-    val result = PriceEquityStep.rewireFirms(firms, 1.0, new Random(42))
+    val result = PriceEquityEconomics.rewireFirms(firms, 1.0, new Random(42))
     result shouldBe theSameInstanceAs(firms)
   }
 
   it should "preserve sector assignment" in {
     val firms  = mkFirmsWithBankrupt(20, 3, sector = 2)
-    val result = PriceEquityStep.rewireFirms(firms, 1.0, new Random(42))
+    val result = PriceEquityEconomics.rewireFirms(firms, 1.0, new Random(42))
     for i <- 0 until 3 do result(i).sector.toInt shouldBe 2
   }
 
   it should "set initialSize on new entrants matching their worker count" in {
     val firms  = mkFirmsWithBankrupt(20, 3)
-    val result = PriceEquityStep.rewireFirms(firms, 1.0, new Random(42))
+    val result = PriceEquityEconomics.rewireFirms(firms, 1.0, new Random(42))
     for i <- 0 until 3 do
       val f = result(i)
       f.initialSize should be > 0
