+++ src/test/scala/com/boombustgroup/amorfati/engine/economics/SigmaDynamicsSpec.scala
@@ -1,4 +1,4 @@
-package com.boombustgroup.amorfati.engine.steps
+package com.boombustgroup.amorfati.engine.economics
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
@@ -10,18 +10,18 @@ class SigmaDynamicsSpec extends AnyFlatSpec with Matchers:
   given SimParams = SimParams.defaults
   private val td  = ComputationBoundary
 
-  "PriceEquityStep.evolveSigmas" should "return unchanged sigmas when lambda=0" in {
+  "PriceEquityEconomics.evolveSigmas" should "return unchanged sigmas when lambda=0" in {
     val current  = Vector(50.0, 10.0, 5.0, 2.0, 1.0, 3.0).map(Sigma(_))
     val base     = Vector(50.0, 10.0, 5.0, 2.0, 1.0, 3.0)
     val adoption = Vector(0.5, 0.3, 0.2, 0.1, 0.0, 0.1)
-    PriceEquityStep.evolveSigmas(current, base, adoption, 0.0, 3.0) shouldBe current
+    PriceEquityEconomics.evolveSigmas(current, base, adoption, 0.0, 3.0) shouldBe current
   }
 
   it should "increase sigma when lambda>0 and adoption>0" in {
     val current  = Vector(Sigma(5.0))
     val base     = Vector(5.0)
     val adoption = Vector(0.5)
-    val result   = PriceEquityStep.evolveSigmas(current, base, adoption, 0.02, 3.0)
+    val result   = PriceEquityEconomics.evolveSigmas(current, base, adoption, 0.02, 3.0)
     td.toDouble(result(0)) should be > 5.0
   }
 
@@ -29,7 +29,7 @@ class SigmaDynamicsSpec extends AnyFlatSpec with Matchers:
     val current  = Vector(Sigma(5.0))
     val base     = Vector(5.0)
     val adoption = Vector(0.0)
-    val result   = PriceEquityStep.evolveSigmas(current, base, adoption, 0.02, 3.0)
+    val result   = PriceEquityEconomics.evolveSigmas(current, base, adoption, 0.02, 3.0)
     td.toDouble(result(0)) shouldBe 5.0
   }
 
@@ -38,7 +38,7 @@ class SigmaDynamicsSpec extends AnyFlatSpec with Matchers:
     val current  = Vector(Sigma(14.99))
     val base     = Vector(5.0)
     val adoption = Vector(1.0)
-    val result   = PriceEquityStep.evolveSigmas(current, base, adoption, 10.0, 3.0)
+    val result   = PriceEquityEconomics.evolveSigmas(current, base, adoption, 10.0, 3.0)
     td.toDouble(result(0)) should be <= 15.0
   }
 
@@ -47,7 +47,7 @@ class SigmaDynamicsSpec extends AnyFlatSpec with Matchers:
     val base     = Vector(5.0)
     // Negative adoption is pathological but ratchet should still hold
     val adoption = Vector(-0.5)
-    val result   = PriceEquityStep.evolveSigmas(current, base, adoption, 0.02, 3.0)
+    val result   = PriceEquityEconomics.evolveSigmas(current, base, adoption, 0.02, 3.0)
     td.toDouble(result(0)) should be >= 5.0
   }
 
@@ -55,7 +55,7 @@ class SigmaDynamicsSpec extends AnyFlatSpec with Matchers:
     val current  = Vector(50.0, 10.0, 5.0, 2.0, 1.0, 3.0).map(Sigma(_))
     val base     = Vector(50.0, 10.0, 5.0, 2.0, 1.0, 3.0)
     val adoption = Vector(0.5, 0.0, 0.3, 0.0, 0.0, 0.2)
-    val result   = PriceEquityStep.evolveSigmas(current, base, adoption, 0.02, 3.0)
+    val result   = PriceEquityEconomics.evolveSigmas(current, base, adoption, 0.02, 3.0)
     // Sectors with adoption > 0 should increase
     td.toDouble(result(0)) should be > 50.0
     td.toDouble(result(2)) should be > 5.0
